core/requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
black
flake8
djangorestframework
markdown
django-filter
djangorestframework-simplejwt